src: remove convinience methods
- These methods are already provided by the cron-parser module.
- Remove addDateConvenienceMethods function from the public API.
- Remove tests.
