+++ lib/schedule.js
@@ -7,7 +7,6 @@
 
 var events = require('events'),
   util = require('util'),
-  increment = require('./increment.js'),
   cronParser = require('cron-parser'),
   lt = require('long-timeout');
 
@@ -270,19 +269,16 @@ function RecurrenceRule(year, month, date, dayOfWeek, hour, minute, second) {
 
 RecurrenceRule.prototype.nextInvocationDate = function(base) {
   base = (base instanceof Date) ? base : (new Date());
-  increment.addDateConvenienceMethods(base);
   if (!this.recurs) {
     return null;
   }
 
   var now = new Date();
-  increment.addDateConvenienceMethods(now);
   if (this.year !== null && (typeof this.year == 'number') && this.year < now.getFullYear()) {
     return null;
   }
 
   var next = new Date(base.getTime());
-  increment.addDateConvenienceMethods(next);
   next.addSecond();
 
   while (true) {
@@ -576,4 +572,3 @@ exports.scheduledJobs = scheduledJobs; // Maybe we should remove this from the p
 exports.cancelJob = cancelJob;
 exports.getJob = getJob;
 exports.getJobs = getJobs;
-exports.addDateConvenienceMethods = increment.addDateConvenienceMethods;
